Move and re-use CollapsibleOutput to hide verbose InstanceHistory fault message + traceback
